Fix AS OF SYSTEM TIME handling
For some reason (not fully debugged), the current AOST handling (where
we supply a timestamp with the system time) causes the query to think
that the database is not found. This only occurs with some timestamp
values, which makes me suspect that it's a substitution issue.
Regardless, the current approach of using a static timestamp is not
ideal, as the workload will break once GC kicks in and removes all data
below that timestamp. This commit chnages the behavior so that AOST
queries are always run 1s in the past.